+++ udata_front/theme/gouvfr/datagouv-components/src/composables/index.ts
@@ -2,5 +2,9 @@ import useOrganizationCertified, { CERTIFIED, PUBLIC_SERVICE } from "./organizat
 import useOwnerName from "./organizations/useOwnerName";
 import useResourceImage from "./resources/useResourceImage";
 import useSchema from "./resources/useSchema";
+import useComponentsForHook from "./useComponentsForHook";
+import useOEmbed from "./useOEmbed";
+import useTabs from "./useTabs";
+import useToast from "./useToast";
 
-export { useOrganizationCertified, useOwnerName, useResourceImage, useSchema, CERTIFIED, PUBLIC_SERVICE };
+export { useComponentsForHook, useOEmbed, useOrganizationCertified, useOwnerName, useResourceImage, useSchema, useTabs, useToast, CERTIFIED, PUBLIC_SERVICE };